+++ wallet/core/src/lib.rs
@@ -6,7 +6,7 @@ pub mod encryption;
 pub mod error;
 pub mod events;
 mod imports;
-pub mod misc;
+pub mod message;
 pub mod result;
 pub mod runtime;
 pub mod secret;
