+++ wallet/core/src/message.rs
