--- wallet/core/src/misc/mod.rs
@@ -1,1 +1,0 @@
-pub mod message_signing;
